95:Created Policies to unables users to edit jobs that they did not create

+++ app/Http/Controllers/JobController.php
@@ -8,9 +8,11 @@ use Illuminate\Http\RedirectResponse;
 use App\Models\Job;
 use Illuminate\Contracts\Support\ValidatedData;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class JobController extends Controller
 {
+    use AuthorizesRequests;
     // @desc Show all job listings
     // @route GET /jobs
     public function index(): View
@@ -82,6 +84,9 @@ class JobController extends Controller
     // @route GET /jobs/{$id}/edit
     public function edit(Job $job): View
     {
+        //check if user is authorized
+        $this->authorize('update', $job);
+        //show the form
         return view('jobs.edit')->with('job', $job);
     }
 
@@ -89,6 +94,9 @@ class JobController extends Controller
     // @route PUT /jobs/{$id}
     public function update(Request $request, Job $job)
     {
+        //check if user is authorized
+        $this->authorize('update', $job);
+
         // Validate the incoming request data
     $validatedData = $request->validate([
         'title' => 'required|string|max:255',
@@ -133,6 +141,9 @@ class JobController extends Controller
     // @route DELETE /jobs{$id}
     public function destroy(Job $job): RedirectResponse
     {
+        //check if user is authorized
+        $this->authorize('delete', $job);
+
         // If logo, then delete it
         if($job->company_logo) {
             Storage::delete('public/logos' . basename($job->company_logo));
